% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cimgConversions.R
\name{array.to.cimg}
\alias{array.to.cimg}
\title{Converts from a raster array to a cimg object}
\usage{
array.to.cimg(x, rm.alpha = TRUE)
}
\arguments{
\item{x}{An image array, i.e. as read in by readPNG.}

\item{rm.alpha}{Logical. Remove the alpha channel?
Note this will "reveal" whatever is hidden behind
the transparent pixels, rather than turn them white.}
}
\value{
A \code{cimg} object.
}
\description{
What it says it does.
}
