% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cimgConversions.R
\name{array_to_cimg}
\alias{array_to_cimg}
\title{Converts from a raster array to a cimg object}
\usage{
array_to_cimg(x, flatten_alpha = TRUE, bg = "white", rm_alpha = TRUE)
}
\arguments{
\item{x}{An image array, i.e. as read in by readPNG.}

\item{flatten_alpha}{Logical. Flatten the alpha channel?}

\item{bg}{Passed to \code{\link[imager]{flatten.alpha}}. Pixel color for
previously transparent pixels.}

\item{rm_alpha}{Logical. Remove the alpha channel?
Note this will "reveal" whatever is hidden behind
the transparent pixels, rather than turn them white.}
}
\value{
A \code{cimg} object.
}
\description{
What it says it does.
}
