% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignPixels.R
\name{assignPixels}
\alias{assignPixels}
\title{Assign a 2D matrix of pixels to specified colors}
\usage{
assignPixels(color.centers, pixel.matrix, adjust.centers = TRUE)
}
\arguments{
\item{color.centers}{Matrix of color centers (rows = colors, columns = channels).}

\item{pixel.matrix}{Matrix of pixel colors (rows = pixels, columns = channels).}

\item{adjust.centers}{Logical. Should the returned color clusters be the
average value of the pixels assigned to that cluster? See details.}
}
\value{
A list of class \code{color.clusters}, containing:
\enumerate{
\item \code{pixel.assignments}: A vector of color center assignments for each pixel.
\item \code{centers}: A matrix of color centers. If \code{adjust.centers = FALSE}, this will be identical to the input \code{color.centers}.
\item \code{sizes}: The number of pixels assigned to each cluster.
}
}
\description{
Assign a 2D matrix of pixels to specified colors
}
\details{
This is a largely internal function called by \code{\link[recolorize]{imposeColors}}
for recoloring an image based on extrinsic colors. If \code{adjust.centers = TRUE},
then after assigning pixels to given colors,
}
\examples{

# RGB extremes (white, black, red, green, blue, yellow, magenta, cyan)
ctrs <- matrix(c(1, 1, 1,
                 0, 0, 0,
                 1, 0, 0,
                 0, 1, 0,
                 0, 0, 1,
                 1, 1, 0,
                 1, 0, 1,
                 0, 1, 1), byrow = TRUE, ncol = 3)

# plot it
recolorize::plotColorPalette(ctrs)

# create a pixel matrix of random colors
pixel.matrix <- matrix(runif(3000), ncol = 3)

# assign pixels
reassigned <- recolorize::assignPixels(ctrs, pixel.matrix, adjust.centers = TRUE)
recolorize::plotColorPalette(reassigned$centers)

# if we turn off adjust.centers, the colors remain the same as the inputs:
keep.centers <- recolorize::assignPixels(ctrs, pixel.matrix, adjust.centers = FALSE)
recolorize::plotColorPalette(keep.centers$centers)

}
