% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brick_to_array.R
\name{brick_to_array}
\alias{brick_to_array}
\title{Convert from a RasterBrick to an array}
\usage{
brick_to_array(raster_brick)
}
\arguments{
\item{raster_brick}{An object of RasterBrick class.}
}
\value{
An image array (probably 1, 3, or 4 channels).
}
\description{
Converts from a RasterBrick to a numeric array. Useful
in going from patternize to recolorize.
}
\details{
This function is provided to convert from the RasterBrick objects provided
by the alignment functions in the patternize package, e.g. \code{alignLan}.
}
