% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cimgConversions.R
\name{cimg.to.array}
\alias{cimg.to.array}
\title{Converts from cimg to raster array}
\usage{
cimg.to.array(x)
}
\arguments{
\item{x}{A \code{cimg} object.}
}
\value{
A 3D array.
}
\description{
What it says it does.
}
