% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cimgConversions.R
\name{cimg_to_array}
\alias{cimg_to_array}
\title{Converts from cimg to raster array}
\usage{
cimg_to_array(x)
}
\arguments{
\item{x}{A \code{cimg} object.}
}
\value{
A 3D array.
}
\description{
What it says it does.
}
