% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recolorize_adjacency.R
\name{classify_recolorize}
\alias{classify_recolorize}
\title{Convert a \code{recolorize} object to a \code{classify} object}
\usage{
classify_recolorize(recolorize_obj, imgname = "")
}
\arguments{
\item{recolorize_obj}{A \code{recolorize} object.}

\item{imgname}{Name of the image (a string).}
}
\value{
A \link[pavo:classify]{pavo::classify} object. The background patch will always
be the first color (patch 1), and will be white by default.
}
\description{
Converts a \link{recolorize} object to a \link[pavo:classify]{pavo::classify} object for
use in pavo.
}
\details{
This is mostly for internal use, and hasn't been tested much.
}
