% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorClusters.R
\name{colorClustersKMeans}
\alias{colorClustersKMeans}
\title{Cluster pixel colors using K-means clustering}
\usage{
colorClustersKMeans(pixel.matrix, n = 10)
}
\arguments{
\item{pixel.matrix}{2D matrix of pixels to classify (rows = pixels, columns =
channels).}

\item{n}{Number of clusters to fit.}
}
\value{
A list with the following elements:
\enumerate{
\item \code{pixel.assignments}: A vector of color center assignments for each pixel.
\item \code{centers}: A matrix of color centers.
\item \code{sizes}: The number of pixels assigned to each cluster.
}
}
\description{
Clusters pixel colors using \code{\link[stats]{kmeans}}.
}
\details{
Called by \code{\link{colorClusters}}. See that documentation for
examples.
}
