% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_expand.R
\name{expand_recolorize}
\alias{expand_recolorize}
\title{Expand aspects of a recolorize object for other functions}
\usage{
expand_recolorize(
  recolorize_obj,
  original_img = FALSE,
  recolored_img = FALSE,
  sizes = FALSE
)
}
\arguments{
\item{recolorize_obj}{A \code{recolorize} object.}

\item{original_img}{Logical. Return original image as numeric array?}

\item{recolored_img}{Logical. Return recolored image as numeric array?}

\item{sizes}{Logical. Return cluster sizes (as number of pixels)?}
}
\value{
A \code{recolorize} object with the indicated additional elements,
as well as the original elements.
}
\description{
Expand aspects of a recolorize object for other functions
}
