% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recluster.R
\name{labelCol}
\alias{labelCol}
\title{Change colors of dendrogram tips}
\usage{
labelCol(x, hex_cols, pch = 20, cex = 2)
}
\arguments{
\item{x}{Leaf of a dendrogram.}

\item{hex_cols}{Hex color codes for colors to change to.}

\item{pch}{The type of point to draw.}

\item{cex}{The size of the point.}
}
\value{
An \code{hclust} object with colored tips.
}
\description{
Internal function for \link{recluster} plotting.
}
