% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medianColors.R
\name{medianColors}
\alias{medianColors}
\title{Change color centers to median color of all pixels assigned to it}
\usage{
medianColors(recolorize_obj, plotting = TRUE)
}
\arguments{
\item{recolorize_obj}{A \code{recolorize} class object.}

\item{plotting}{Logical. Plot results?}
}
\value{
A \code{recolorize} object, with median colors instead of average colors
in the \code{centers} attribute.
}
\description{
By default, recolorize sets the centers of each color patch to the average
(mean) color of all pixels assigned to it. This can sometimes result in colors
that look washed out, especially in cases where a region is very shiny (e.g.
black with white reflective highlights will average to grey). In these cases,
switching to median colors may be either more accurate or more visually
pleasing.
}
