% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixelAssignMatrix.R
\name{pixelAssignMatrix}
\alias{pixelAssignMatrix}
\title{Make pixel assignment matrix for recoloring}
\usage{
pixelAssignMatrix(bg_indexed, color_clusters)
}
\arguments{
\item{bg_indexed}{An object returned by \code{\link{backgroundIndex}}.}

\item{color_clusters}{An object returned by \code{\link{colorClusters}}.}
}
\value{
A matrix of pixel color assignments (\code{pixel_assignments})
and a corresponding dataframe of color centers (\code{centers}).
}
\description{
Internal function. Generates a sort of 'paint-by-numbers' matrix, where each
cell is the index of the color in the color centers matrix to which that
pixel is assigned. An index of 0 indicates a background pixel.
}
