% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recolorizeVector.R
\name{plot.recolorizeVector}
\alias{plot.recolorizeVector}
\title{Plot a \code{recolorizeVector} object}
\usage{
\method{plot}{recolorizeVector}(x, ...)
}
\arguments{
\item{x}{Object returned by \link{recolorizeVector}.}

\item{...}{Further arguments passed to \link[graphics:plot.default]{graphics::plot}.}
}
\description{
Plots an object generated by \link{recolorizeVector}.
}
