% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRecolorized.R
\name{plotRecolorized}
\alias{plotRecolorized}
\title{Plot recolorized image results}
\usage{
plotRecolorized(
  recolored_img,
  original_img,
  plot_original = TRUE,
  centers,
  horiz = FALSE,
  cex_text = 2,
  sizes = NULL
)
}
\arguments{
\item{recolored_img}{The recolored image, as a 3D RGB array.}

\item{original_img}{The original image, as a 3D RGB array.}

\item{plot_original}{Logical. Plot the original image for comparison?}

\item{centers}{The colors to plot in the color palette, as a numeric
matrix (rows = colors, columns = channels).}

\item{horiz}{Logical. Should plots be stacked vertically or horizontally?}

\item{cex_text, sizes}{Plotting parameters passed to
\code{\link{recolorize}{plotColorPalette}}.}
}
\description{
Plots a side-by-side comparison of an original image and its recolorized
version, plus the color palette used for recoloring. Called by
\code{\link[recolorize]{imposeColors}} and
\code{\link[recolorize]{recolorize}}.
}
\examples{
corbetti <- system.file("extdata/corbetti.png",
                         package = "recolorize")

corbetti_recolor <- recolorize(corbetti, method = "hist",
                                         bins = 2, plotting = FALSE)

# full color palette
plotRecolorized(recolored_img = corbetti_recolor$recolored_img,
                original_img = corbetti_recolor$original_img,
                centers = corbetti_recolor$centers)

# scaled color palette
plotRecolorized(recolored_img = corbetti_recolor$recolored_img,
                original_img = corbetti_recolor$original_img,
                centers = corbetti_recolor$centers,
                sizes = corbetti_recolor$sizes)

}
