% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_array_conversions.R
\name{raster_to_array}
\alias{raster_to_array}
\title{Convert from a (small-r) raster object to an RGB array}
\usage{
raster_to_array(raster_obj, alpha = TRUE)
}
\arguments{
\item{raster_obj}{A matrix of hex codes as output by \link[grDevices:as.raster]{grDevices::as.raster}.}

\item{alpha}{Logical. If there is an alpha channel, retain it in the array?}
}
\value{
A numeric RGB array (0-1 range).
}
\description{
Recreates the original numeric array from a \code{raster} object created
by \link[grDevices:as.raster]{grDevices::as.raster}. Not to be confused with the \verb{Raster*} classes
used by the \code{raster} package.
}
