% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recolorize-package.R
\docType{package}
\name{recolorize-package}
\alias{recolorize-package}
\title{recolorize: Color-Based Image Segmentation}
\description{
Automatic, semi-automatic, and manual functions for generating color maps from images. The idea is to simplify the colors of an image according to a metric that is useful for the user, using deterministic methods whenever possible. Many images will be clustered well using the out-of-the-box functions, but the package also includes a toolbox of functions for making manual adjustments (layer merging/isolation, blurring, fitting to provided color clusters or those from another image, etc). Also includes export methods for other color/pattern analysis packages (pavo, patternize, colordistance).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://hiweller.github.io/recolorize/}
  \item \url{https://github.com/hiweller/recolorize}
  \item Report bugs at \url{https://github.com/hiweller/recolorize/issues}
}

}
\author{
\strong{Maintainer}: Hannah Weller \email{hannahiweller@gmail.com} (\href{https://orcid.org/0000-0002-5252-4282}{ORCID})

}
\keyword{internal}
