% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recolorize.R
\name{recolorize}
\alias{recolorize}
\title{Simplify the colors of an image}
\usage{
recolorize(
  img.path,
  method = "kmeans",
  bins = 3,
  n = 10,
  lower = NULL,
  upper = NULL,
  transparent = TRUE,
  resize = NULL,
  rotate = NULL,
  plotting = TRUE,
  horiz = TRUE,
  cex.text = 1.5,
  scale.palette = TRUE
)
}
\arguments{
\item{img.path}{Path to the image. Must be a character vector.}

\item{lower, upper}{RGB triplet ranges for setting a bounding box of pixels to
mask. See details.}

\item{transparent}{Logical. Treat transparent pixels as background? Requires
an alpha channel (PNG).}

\item{resize}{A value between 0 and 1 for resizing the image (ex. \code{resize = 0.5} will reduce image size by 50\%). Recommended for large images as it can
speed up analysis considerably. See details.}

\item{rotate}{Degrees to rotate the image clockwise.}

\item{plotting}{Logical. Plot recolored image & color palette?}

\item{horiz}{Logical for plotting. Plot output image and color palette side
by side (\code{TRUE}) or stacked vertically (\code{FALSE})?}

\item{cex.text}{If \code{plotting = TRUE} and \code{scale.palette = FALSE}, size of
text to display on the color palette numbers.}

\item{scale.palette}{Logical. If plotting, plot colors in the color palette
proportional to the size of each cluster?}

\item{color.centers}{Colors to map to, as an n x 3 matrix (rows = colors,
columns = channels).}

\item{adjust.centers}{Logical. After pixel assignment, should the returned
colors be the average color of the pixels assigned to that cluster, or the
original colors?}
}
\value{
A list with the following attributes:
\enumerate{
\item \code{original.img}: The original image, as a 3D array.
\item \code{recolored.img}: The recolored image, as a 3D array.
\item \code{color.space}: The associated color space. Currently only RGB.
\item \code{centers}: A matrix of color centers. If \code{adjust.centers = FALSE}, this will be identical to the input \code{color.centers}.
\item \code{sizes}: The number of pixels assigned to each color cluster.
\item \code{pixel.assignments}: A vector of color center assignments for each pixel.
}
}
\description{
Takes an image and a set of color centers, and assigns each pixel to the most
similar provided color. Useful for producing a set of images with identical colors.
}
