% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitByColor.R
\name{splitByColor}
\alias{splitByColor}
\title{Split color clusters in a recolorize object into layers}
\usage{
splitByColor(recolorize_obj, layers = "all", plot_method = "overlay")
}
\arguments{
\item{recolorize_obj}{A recolorize object from \code{\link{recolorize}},
\code{\link{recluster}}, or \code{\link{imposeColors}}.}

\item{layers}{Either \code{"all"} or a numeric vector of which color centers to
return.}

\item{plot_method}{Plotting method for plotting the color layers. Options
are\code{"overlay"}, \code{"binary"}, \code{"colormask"}, or \code{"none"}.}
}
\value{
A list of binary matrices (1/white = color presence, 0/black = color
absence), one per color center.
}
\description{
Separates color clusters from a \code{\link{recolorize}},
\code{\link{recluster}}, or \code{\link{imposeColors}} object
into binary masks.
}
\examples{
# get original fit
corbetti <- system.file("extdata/corbetti.png", package = "recolorize")
recolored_corbetti <- recolorize::recolorize(corbetti, plotting = TRUE)

# make a layout
layout(matrix(c(1, 1:9), nrow = 2))

# plot original
plotImageArray(recolored_corbetti$original_img)

# plot layers
corbetti_layers <- splitByColor(recolored_corbetti, plot_method = "over")

# plot one of the binary maps
layout(matrix(1:2, nrow = 1))
plotImageArray(corbetti_layers[[2]], main = "original layer")

# make an imager pixset
px <- imager::as.cimg(corbetti_layers[[2]] == 1)
px_clean <- imager::clean(px, 3)

# convert back to an image for plotting
plotImageArray(recolorize:::cimg_to_array(px_clean), main = "cleaned layer")

}
