% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wernerColor.R
\docType{data}
\name{werner}
\alias{werner}
\title{Werner's nomenclature of colors}
\format{
A data frame with 110 rows and 13 variables:
\describe{
\item{index}{The color index.}
\item{family}{The broad color category (white, red, etc).}
\item{name}{The original color name.}
\item{hex}{Color hex code.}
}
}
\source{
\url{https://www.c82.net/werner/#colors}
}
\usage{
werner
}
\description{
A table of the 110 colors described in "Werner's Nomenclature of Colors", the
1821 color reference by Patrick Syme (building on work by Abraham Gottlob
Werner), notably used by Charles Darwin. Colors represent the average pixel
color of each scanned swatch.
}
\keyword{datasets}
