% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cimgConversions.R
\name{array.to.cimg}
\alias{array.to.cimg}
\title{Converts from a raster array to a cimg object}
\usage{
array.to.cimg(x)
}
\arguments{
\item{x}{An image array, i.e. as read in by readPNG.}
}
\value{
A \code{cimg} object.
}
\description{
What it says it does.
}
